+++ examples/pub_sub.js
@@ -30,10 +30,9 @@ client1.on("message", function (channel, message) {
     }
 });
 
-client1.incr("did a thing");
-
 client1.on("ready", function () {
     // if you need auth, do it here
+    client1.incr("did a thing");
     client1.subscribe("a nice channel", "another one");
 });